Revert "feat: partial upgrade Apache PDFBox from 2.0.34 to 3.0.5"
This reverts commit 82e70f9a65a1a4a72a9bd16a50d59118f5beb2d0.

+++ pdf-as-pdfbox-2/src/main/java/at/knowcenter/wag/egov/egiz/pdfbox2/pdf/operator/path/construction/CurveToReplicateFinalPoint.java
@@ -85,8 +85,8 @@ public class CurveToReplicateFinalPoint extends PathConstructionOperatorProcesso
 			COSNumber x3 = (COSNumber) operands.get(2);
 			COSNumber y3 = (COSNumber) operands.get(3);
 
-			Point2D p1 = transform(x1.floatValue(), y1.floatValue());
-			Point2D p3 = transform(x3.floatValue(), y3.floatValue());
+			Point2D p1 = transform(x1.doubleValue(), y1.doubleValue());
+			Point2D p3 = transform(x3.doubleValue(), y3.doubleValue());
 
 			pdfPage.getCurrentPath().curveTo(
 					(float) p1.getX(), (float) p1.getY(),
